+++ im_flutter_sdk_ios/ios/im_flutter_sdk_ios/Sources/im_flutter_sdk_ios/PushManagerWrapper.m
@@ -86,6 +86,13 @@
     else if([bindDeviceToken isEqualToString:call.method]) {
         [self bindDeviceToken:call.arguments channelName:call.method result:result];
     }
+    // 5.0.0
+    else if([bindPushKitToken isEqualToString:call.method]) {
+        [self bindPushKitToken:call.arguments channelName:call.method result:result];
+    }
+    else if([unbindPushKitToken isEqualToString:call.method]) {
+        [self unbindPushKitToken:call.arguments channelName:call.method result:result];
+    }
     else{
         [super handleMethodCall:call result:result];
     }
@@ -330,8 +337,8 @@
             channelName:(NSString *)aChannelName
                  result:(FlutterResult)result {
     __weak typeof(self) weakSelf = self;
-    NSString *notifierName = param[@"notifierName"];
-    EMClient.sharedClient.options.apnsCertName = notifierName;
+    // The APNs certificate name is not taken from the request: it is an initialization option
+    // (ChatOptions.apnsCertName -> EMOptions.apnsCertName) and cannot change at runtime.
     NSString *deviceToken = param[@"deviceToken"];
     dispatch_async(dispatch_get_global_queue(DISPATCH_QUEUE_PRIORITY_DEFAULT, 0), ^{
         [EMClient.sharedClient bindFCMToken:deviceToken completion:^(EMError * _Nullable aError) {
@@ -341,6 +348,39 @@
                                object:nil];
         }];
     });
+}
+
+#pragma mark - 5.0.0
+
+- (void)bindPushKitToken:(NSDictionary *)param
+             channelName:(NSString *)aChannelName
+                  result:(FlutterResult)result {
+    __weak typeof(self) weakSelf = self;
+    // The PushKit certificate name is not taken from the request: it is an initialization option
+    // (ChatOptions.pushKitCertName -> EMOptions.pushKitCertName) and cannot change at runtime.
+    // An empty certificate name makes the native call fail with EMErrorUserIllegalArgument.
+    NSString *deviceToken = param[@"deviceToken"];
+    // The native _extractTokenFromRawData: accepts both NSData and NSString, which is the
+    // same way the APNs route above passes its token through bindFCMToken:.
+    [EMClient.sharedClient registerPushKitToken:(NSData *)deviceToken
+                                     completion:^(EMError * _Nullable aError) {
+        [weakSelf wrapperCallBack:result
+                      channelName:aChannelName
+                            error:aError
+                           object:nil];
+    }];
+}
+
+- (void)unbindPushKitToken:(NSDictionary *)param
+               channelName:(NSString *)aChannelName
+                    result:(FlutterResult)result {
+    __weak typeof(self) weakSelf = self;
+    [EMClient.sharedClient unRegisterPushKitTokenWithCompletion:^(EMError * _Nullable aError) {
+        [weakSelf wrapperCallBack:result
+                      channelName:aChannelName
+                            error:aError
+                           object:nil];
+    }];
 }
 
 @end